+++ gazebo/sensors/MultiCameraSensor.cc
@@ -20,6 +20,8 @@
   #include <Winsock2.h>
 #endif
 
+#include <boost/algorithm/string.hpp>
+#include <boost/bind.hpp>
 #include <ignition/math/Pose3.hh>
 
 #include "gazebo/common/Exception.hh"